Add an aubiolevel CLI tool to compute the sound pressure level
Aubio enables users to compute the sound pressure level of a buffer
using miscutils.h, but there is no way to get that information from CLI
tools right now.

This adds support for a standalone CLI tool so that extracting sound
pressure level information does not require building a program against
the library.